Find parameters wizard

+++ plugin/src/mosaic/region_competition/Settings.java
@@ -61,7 +61,7 @@ public class Settings implements Serializable
 	public int m_BubblesDispl = 10;
 	
 	public int l_BubblesRadius = 5;
-	public int l_Sigma = 2;
+	public double l_Sigma = 2;
 	public double l_Tolerance = 0.005;
 	public int l_RegionTolerance = 4;
 	public double l_BoxRatio = 0.8;
